chore: move management/controller replication slot logic in a specific file
